Changed the way events are handled, removed some non-working functions and renamed others
Added ability to use slash commands and regular commands at the same time without first one being overwritten by the second.

Removed "modifyCurrentUserNick" as it is in "modifyCurrentMemberNick"

Renamed: Refresh Global Application Commands and Refresh Application Commands to delete and changed it to actually delete the commands as updating is as simple as overwriting existing one.

Added replyToInteraction method to reply to slash commands.